Create config subdirectory
This is a preparational step for implementing an internal configuration
format.

CMK-14802
